Add displayName to React class
By setting an explicit `displayName`, the InlineEdit component will be correctly named when viewed from the React Dev Tools. I think this is normally done implicitly when [in a `.jsx` file](https://facebook.github.io/react/docs/jsx-in-depth.html#the-transform) or when using ES6 classes, but as this component doesn't fall into either category it just shows up as "<No display name [...] />" in the dev tools currently.